Renamed query.py to query_and_mutations.py
